+++ lib/builder.dart
@@ -1,0 +1,3 @@
+library;
+
+export 'src/ql_gen_builder.dart' show qlGenBuilder;